Clarify how to develop on Windows
There is an issue with `doc8` on windows in that it is not very understanding of the different line endings.

Git on windows, at least in the default configuration, will convert between \r\n line endings and \n line endings, seemlessly.

This makes doc8 break. We could have it --ignore D004 --ignore D002, but that would  not be entirely correct.