Fix broken CI (#1006)
The jobs are currently failing due to a mismatch in Python versions.

The reason is that we implicitly use the latest `sphinxdoc/sphinx` image while relying on a specific version of Python being available. It might be a good idea to pin the docker image versions to avoid surprises like this in the future.